src/main/java/top/aranlzh/sell/service/impl/ProductServiceImpl.java
package top.aranlzh.sell.service.impl;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;
import top.aranlzh.sell.dataobject.ProductInfo;
import top.aranlzh.sell.dto.CartDTO;
import top.aranlzh.sell.enums.ProductStatusEnum;
import top.aranlzh.sell.enums.ResultEnum;
import top.aranlzh.sell.exception.SellException;
import top.aranlzh.sell.repository.ProductInfoRepository;
import top.aranlzh.sell.service.ProductService;

import javax.transaction.Transactional;
import java.util.List;

/**
 * @author Aranlzh
 * @create 2019-03-12 0:07
 * @desc
 **/
@Service
public class ProductServiceImpl implements ProductService {

    @Autowired
    private ProductInfoRepository repository;

    @Override
    public ProductInfo findOne(String productId) {
        return repository.findOne(productId);
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return repository.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return repository.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return repository.save(productInfo);
    }

    @Transactional
    public void increaseStock(List<CartDTO> cartDTOList) {
        for (CartDTO cartDTO: cartDTOList) {
            ProductInfo productInfo = repository.findOne(cartDTO.getProductId());
            if (productInfo == null) {
                throw new SellException(ResultEnum.PRODUCT_NOT_EXIST);
            }
            Integer result = productInfo.getProductStock() + cartDTO.getProductQuantity();
            productInfo.setProductStock(result);

            repository.save(productInfo);
        }

    }

    @Transactional
    public void decreaseStock(List<CartDTO> cartDTOList) {
        for (CartDTO cartDTO: cartDTOList) {
            ProductInfo productInfo = repository.findOne(cartDTO.getProductId());
            if (productInfo == null) {
                throw new SellException(ResultEnum.PRODUCT_NOT_EXIST);
            }

            // 计算订单后剩余库存
            Integer result = productInfo.getProductStock() - cartDTO.getProductQuantity();
            // 库存不足的时候
            if (result < 0) {
                throw new SellException(ResultEnum.PRODUCT_STOCK_ERROR);
            }
            // 保存新库存量
            productInfo.setProductStock(result);
            // 刚好卖完的话，下架商品
            if (result == 0 ){
                productInfo.setProductStatus(ProductStatusEnum.DOWN.getCode());
            }
            repository.save(productInfo);
        }
    }

    public ProductInfo onSale(String productId) {
        ProductInfo productInfo = repository.findOne(productId);
        if (productInfo == null) {
            throw new SellException(ResultEnum.PRODUCT_NOT_EXIST);
        }
        if (productInfo.getProductStatusEnum() == ProductStatusEnum.UP) {
            throw new SellException(ResultEnum.PRODUCT_STATUS_ERROR);
        }

        //更新
        productInfo.setProductStatus(ProductStatusEnum.UP.getCode());
        return repository.save(productInfo);
    }

    public ProductInfo offSale(String productId) {
        ProductInfo productInfo = repository.findOne(productId);
        if (productInfo == null) {
            throw new SellException(ResultEnum.PRODUCT_NOT_EXIST);
        }
        if (productInfo.getProductStatusEnum() == ProductStatusEnum.DOWN) {
            throw new SellException(ResultEnum.PRODUCT_STATUS_ERROR);
        }

        //更新
        productInfo.setProductStatus(ProductStatusEnum.DOWN.getCode());
        return repository.save(productInfo);
    }

    @Override
    public Page<ProductInfo> findProductByKey(String productNameKey, Pageable pageable) {
        return repository.findByProductNameLike("%"+productNameKey+"%",pageable);
    }

    @Override
    public List<ProductInfo> findUpProductByKey(String productNameKey) {
        return repository.findByProductStatusAndProductNameLike(ProductStatusEnum.UP.getCode(),"%"+productNameKey+"%");
    }
}
